サイト全体のリファクタリングを実施しました。
