+++ src/components/Navbar.tsx
@@ -1,15 +1,17 @@
 import Link from 'next/link';
 import Image from 'next/image';
+import '../styles/globals.css';
 import DescriptionIcon from '@mui/icons-material/Description';
+import AssignmentIcon from '@mui/icons-material/Assignment';
+import WorkIcon from '@mui/icons-material/Work';
 
 const Navbar = () => {
   return (
     <header className="flex items-center bg-gray-800 text-white p-4">
       <div className="container mx-auto flex justify-between items-center">
-        {/* ロゴ部分 */}
-    <DescriptionIcon />
         <div className="flex items-center">
-          <Image src="/logo.png" alt="Logo" width={50} height={50} />
+          {/* <Image src="/logo.png" alt="Logo" width={50} height={50} /> */}
+          <DescriptionIcon width={50} height={50} />
           <span className="font-bold text-xl ml-2">My Portfolio</span>
         </div>
         <nav>
@@ -29,9 +31,6 @@ const Navbar = () => {
                 <span className="hover:underline cursor-pointer">Contact</span>
               </Link>
             </li>
-            {/* <li onClick={() => handleNavigation('/contact')}>
-              <a className="hover:underline cursor-pointer">Contact</a>
-            </li> */}
           </ul>
         </nav>
       </div>
